Функции перенесены в один файл utils.js
