--- src/utils/convertDuration.js
@@ -1,5 +1,0 @@
-export const convertDuration = (movie) => {
-    const hours = Math.floor(movie.duration / 60)
-    const minutes = movie.duration % 60
-    return `${hours}ч ${minutes}мин`;
-}
